// api/consolidated/models/GetFormEntriesResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { FormEntry } from './FormEntry';

export type GetFormEntriesResponse = {
    entries?: Array<FormEntry> | null;
    total?: number;
    page?: number;
    size?: number;
};