// api/consolidated/models/LookupValue.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type LookupValue = {
    type?: string | null;
    groupName?: string | null;
    key?: string | null;
    display?: string | null;
    icon?: string | null;
    value?: string | null;
    order?: number;
};