// api/consolidated/models/ReferenceLookupSearchResult.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { ReferenceLookup } from './ReferenceLookup';

export type ReferenceLookupSearchResult = {
    items?: Array<ReferenceLookup> | null;
    page?: number;
    size?: number;
    total?: number;
    errors?: Array<string> | null;
    queryUsed?: string | null;
    indexUsed?: string | null;
};